+++ Release_Notes.md
@@ -28,6 +28,7 @@ Marked AbstractDatabaseFunctions and AbstractBaseModel as Obsolete
     - MGroupProfile                         to AAddedUpdated
     - MNameValuePair                        to AAddedUpdated
     - MNameValuePairDetails                 to AAddedUpdated
+    - MRegistrationInformation              to AAddedUpdated
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure